LayoutTests/storage/indexeddb/pending-version-change-on-exit.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<script>

description("No crashes when there are event listeners in a worker on shutdown");

if (self.location.search == "?second") {
  testPassed("Didn't crash!");
  finishJSTest();
} else {
  evalAndLog("gotVersionChangeEvent = false");
  indexedDBTest(prepareDatabase, startTheWorker);
}

function prepareDatabase()
{
  debug("inside prepareDatabase");
  preamble();
  db = event.target.result;
  db.onversionchange = function(e) {
    evalAndLog("gotVersionChangeEvent = true");
    debug("Got a version change event, e.oldVersion = " + e.oldVersion +
          " e.newVersion = " + e.newVersion);
  };
}

function startTheWorker() {
  var worker = startWorker("resources/pending-version-change-on-exit.js?" + encodeURIComponent(dbname));
  realFinishJSTest = finishJSTest;
  worker.onerror = function(e) {
    testFailed(e.message);
    realFinishJSTest();
  };
  finishJSTest = function() {
    shouldBeTrue("gotVersionChangeEvent");
    if (gotVersionChangeEvent)
      self.location = self.location + "?second";
    else
      realFinishJSTest();
  }
}
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
